Создал файл  picture.js, в котором добавил все сгенерированные фотографии на страницу
