+++ js/main.js
@@ -1,1 +1,2 @@
 import {createPhotos} from './data.js';
+import './picturs.js';
